streamlit_app/requirements.txt
streamlit==1.37.1
pandas==2.2.2
google-cloud-bigquery==3.25.0
pyarrow==16.1.0
db-dtypes==1.2.0